Move internal adapter into pkg subpackage
Rename swanlab/sdk/internal/adapter to swanlab/sdk/internal/pkg/adapter and update all imports accordingly. Adjust imports in context/__init__.py and run/record_builder.py to reference swanlab.sdk.internal.pkg.adapter (and pkg.fs), and update unit tests to the new test path. This reorganizes the internal adapter modules under the pkg namespace.
